requirements/base.txt
django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
psycopg2==2.9.9
python-dotenv==1.0.0
django-filter==23.3
# django-phonenumber-field
# phonenumbers
# Pillow
# django-mptt